+++ RELEASE_NOTES.md
@@ -2,6 +2,11 @@
 
 ## XMLUnit.NET 2.11.2 - /not released, yet/
 
+* `IsDateTimePlaceholderHandler` now supports an optional second argument specifying a
+  different CultureInfo than the invariant culture used by default as a BCP 47 language tag.
+
+  based on PR [xmlunit/#335](https://github.com/xmlunit/xmlunit/pull/335) by [@jmestwa-coder](https://github.com/jmestwa-coder)
+
 ## XMLUnit.NET 2.11.1 - /Released 2025-05-19/
 
 * placeholders can now also be used inside of the local part of `xsi:type` attributes.